+++ setup.cfg
@@ -20,5 +20,4 @@ tests_require = pytest
 test = pytest
 
 [tool:pytest]
-addopts = --verbose
 testpaths = tests
