+++ tests/test_bootstrap.py
@@ -1,19 +1,21 @@
+import pytest
 from flask import current_app
 
 
-class TestBootstrap(object):
-    def test_extension_init(self, client, bootstrap):
+@pytest.mark.usefixtures("client")
+class TestBootstrap:
+    def test_extension_init(self, bootstrap):
         assert 'bootstrap' in current_app.extensions
 
-    def test_load_css(self, client, bootstrap):
+    def test_load_css(self, bootstrap):
         rv = bootstrap.load_css()
         assert 'bootstrap.min.css' in rv
 
-    def test_load_js(self, client, bootstrap):
+    def test_load_js(self, bootstrap):
         rv = bootstrap.load_js()
         assert 'bootstrap.min.js' in rv
 
-    def test_local_resources(self, client, bootstrap):
+    def test_local_resources(self, bootstrap, client):
         current_app.config['BOOTSTRAP_SERVE_LOCAL'] = True
 
         response = client.get('/')
@@ -37,7 +39,7 @@ class TestBootstrap(object):
         assert 'https://cdn.jsdelivr.net/npm/bootstrap' not in css_rv
         assert 'https://cdn.jsdelivr.net/npm/bootstrap' not in js_rv
 
-    def test_cdn_resources(self, client, bootstrap):
+    def test_cdn_resources(self, bootstrap, client):
         current_app.config['BOOTSTRAP_SERVE_LOCAL'] = False
 
         response = client.get('/')
@@ -53,26 +55,3 @@ class TestBootstrap(object):
         assert '/bootstrap/static/js/bootstrap.min.js' not in js_rv
         assert 'https://cdn.jsdelivr.net/npm/bootstrap' in css_rv
         assert 'https://cdn.jsdelivr.net/npm/bootstrap' in js_rv
-
-    def test_bootswatch_local(self, client, bootswatch_themes):
-        current_app.config['BOOTSTRAP_SERVE_LOCAL'] = True
-
-        for theme in bootswatch_themes:
-            current_app.config['BOOTSTRAP_BOOTSWATCH_THEME'] = theme
-            data = client.get('/').get_data(as_text=True)
-            assert 'https://cdn.jsdelivr.net/npm/bootswatch' not in data
-            assert 'swatch/%s/bootstrap.min.css' % theme in data
-            with client.get('/bootstrap/static/css/swatch/%s/bootstrap.min.css' % theme) as css_response:
-                assert css_response.status_code != 404
-
-    def test_bootswatch_cdn(self, client, bootstrap, bootswatch_themes):
-        current_app.config['BOOTSTRAP_SERVE_LOCAL'] = False
-
-        for theme in bootswatch_themes:
-            current_app.config['BOOTSTRAP_BOOTSWATCH_THEME'] = theme
-            data = client.get('/').get_data(as_text=True)
-            assert 'https://cdn.jsdelivr.net/npm/bootswatch' in data
-            assert 'dist/%s/bootstrap.min.css' % theme in data
-            css_rv = bootstrap.load_css()
-            assert '/bootstrap/static/css/swatch/%s/bootstrap.min.css' % theme not in data
-            assert 'https://cdn.jsdelivr.net/npm/bootswatch' in css_rv
