+++ tox.ini
@@ -26,6 +26,6 @@ commands =
 [testenv:flake8]
 basepython = python2.7
 commands =
-    flake8 flask_bootstrap test_bootstrap_flask.py
+    flake8 flask_bootstrap tests
 deps =
     flake8
